Tie App() state to header component buttons
The buttons of Header component now successfully change state in
the parent component App(). Using a series of ternery operators,
the buttons elements of Header conditionally get a className of
'active' or 'not-active' and an onClick of props.handleClick or
null depending on which page is selected in the parent component
state.