Complete all form inputs
All form components have inputs that are built and successfully
linked with state of grandparent component (App.js) via useContext.
This state in App.js can also be read as expected in the Resume.js
component to populate the web page.

Next I will attempt to organize my form data as nested objects to
ensure this is possible and to clean up codebase.