Add multiple jobs to Prof component
ProfessionalForm.js component now has feilds for multiple jobs.
This updates several different nested objects inside App.js state.
This is accomplished by tying all input ids to the jobs they
represesnt. Then handleChange now takes the id out of the event
target and uses that to place the data in the right object in state.

Errors were overcome by making a copy of the nested obj in state,
updating the copy, and then placing the copy back inside the
setFormData by the code [`${id}`] : jobNew.

"jobNew" is the variable containing the updated copy of the
nested obj.