Add invalidateField function
invalidatedField() takes the name of a field property as a string
and sets that property in the root state object from true to false.

Currently there is no way to set this back to true, but the code is
otherwise functioning as intended and updating the firstNameValid
state produces the desired visual feedback to the user.